Restructure test directory to match project
