+++ openff/system/tests/interop/test_external.py
@@ -9,7 +9,7 @@ from simtk.unit import nanometer as nm
 
 from openff.system.components.mdtraj import OFFBioTop
 from openff.system.stubs import ForceField
-from openff.system.tests.base_test import BaseTest
+from openff.system.tests import BaseTest
 from openff.system.tests.energy_tests.openmm import (
     _get_openmm_energies,
     get_openmm_energies,
