--- openff/system/tests/test_matrix_representations.py
@@ -1,65 +1,0 @@
-import numpy as np
-import pytest
-from openff.utilities.testing import skip_if_missing
-
-from openff.system.components.smirnoff import (
-    SMIRNOFFAngleHandler,
-    SMIRNOFFBondHandler,
-    SMIRNOFFvdWHandler,
-)
-from openff.system.tests.base_test import BaseTest
-from openff.system.utils import get_test_file_path
-
-
-@skip_if_missing("jax")
-class TestMatrixRepresentations(BaseTest):
-    @pytest.mark.parametrize(
-        "handler_name,n_ff_terms,n_sys_terms",
-        [("vdW", 10, 72), ("Bonds", 8, 64), ("Angles", 6, 104)],
-    )
-    def test_to_force_field_to_system_parameters(
-        self, parsley, ethanol_top, handler_name, n_ff_terms, n_sys_terms
-    ):
-        import jax
-
-        if handler_name == "Bonds":
-            handler, _ = SMIRNOFFBondHandler.from_toolkit(
-                bond_handler=parsley["Bonds"],
-                topology=ethanol_top,
-                constraint_handler=None,
-            )
-        elif handler_name == "Angles":
-            handler = SMIRNOFFAngleHandler.from_toolkit(
-                parameter_handler=parsley[handler_name],
-                topology=ethanol_top,
-            )
-        elif handler_name == "vdW":
-            handler = SMIRNOFFvdWHandler._from_toolkit(
-                parameter_handler=parsley[handler_name],
-                topology=ethanol_top,
-            )
-
-        p = handler.get_force_field_parameters()
-
-        assert isinstance(p, jax.interpreters.xla.DeviceArray)
-        assert np.prod(p.shape) == n_ff_terms
-
-        q = handler.get_system_parameters()
-
-        assert isinstance(q, jax.interpreters.xla.DeviceArray)
-        assert np.prod(q.shape) == n_sys_terms
-
-        assert jax.numpy.allclose(q, handler.parametrize(p))
-
-        param_matrix = handler.get_param_matrix()
-
-        ref_file = get_test_file_path(f"ethanol_param_{handler_name.lower()}.npy")
-        ref = jax.numpy.load(ref_file)
-
-        assert jax.numpy.allclose(ref, param_matrix)
-
-        # TODO: Update with other handlers that can safely be assumed to follow 1:1 slot:smirks mapping
-        if handler_name in ["vdW", "Bonds", "Angles"]:
-            assert np.allclose(
-                np.sum(param_matrix, axis=1), np.ones(param_matrix.shape[0])
-            )
